requirements.txt
mysql-connector-python==8.1.0
numpy==1.25.1
python-dotenv==1.0.0
pandas==2.0.3
selenium==4.10.0
undetected-chromedriver==3.5.0
webdriver-manager==3.8.2
thefuzz==0.19.0